Allow entering non-integer BPM tempos
Some people have mentioned it as being useful,
though the added BPM tempo precision is actually mostly useless
because engine tempos must be integers.
Writing non-integer BPM tempos in a tempo-change effect is tricky,
but may be possible through multi-byte effects (whether through a
dedicated effect ID for 2-byte tempo change, or an optional "extension
byte" effect ID entered next to single-byte effects).

+++ src/edit/edit_doc.cpp
@@ -55,9 +55,9 @@ static double & get_tempo_mut(doc::Document & document) {
     return document.sequencer_options.target_tempo;
 };
 
-EditBox set_tempo(int tempo) {
+EditBox set_tempo(double tempo) {
     return make_command(Setter<double, get_tempo_mut, ModifiedFlags::TargetTempo> (
-        (double) tempo
+        tempo
     ));
 }
